Move port utility out of endpoints controller
